pyproject.toml
[tool.poetry]
name = "coursework-5"
version = "0.1.0"
description = "Проект по работе с API HeadHunter и базой данных PostgreSQL"
authors = ["BroneKot"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
